:white_check_mark: Add subscription product register service test
