+++ src/main/java/kr/bb/product/domain/product/application/usecase/ProductQueryUseCase.java
@@ -1,8 +1,19 @@
 package kr.bb.product.domain.product.application.usecase;
 
 import kr.bb.product.domain.product.entity.ProductCommand;
+import kr.bb.product.domain.product.entity.ProductCommand.ProductList;
+import org.springframework.data.domain.Pageable;
 
 public interface ProductQueryUseCase {
-    void updateProductSaleStatus(String productId, ProductCommand.ProductUpdate productRequestData);
-    void createProduct(ProductCommand.ProductRegister productRequestData);
+  ProductList getProductsByCategory(Long userId, Long categoryId, Pageable pageable);
+
+  ProductList getProductsByCategory(Long categoryId, Pageable pageable);
+
+  ProductList getProductsByTag(Long tagId, Pageable pageable);
+
+  ProductList getProductsByTag(Long userId, Long tagId, Pageable pageable);
+
+  ProductCommand.ProductDetail getProductDetail(Long userId, String productId);
+
+  ProductCommand.ProductDetail getProductDetail(String productId);
 }
